chore(txlogic): add new tx and make vout vin utils apis
